requirements.txt
pymupdf
openai
langchain
langchain-openai
faiss-cpu
python-dotenv
langchain_community